python/utopia/requirement.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.13.0
idna==3.3
Pillow==9.1.1
psycopg==3.0.15
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toposort==1.7
tzdata==2022.1
urllib3==1.26.9
